broke app into 3 seperate components. Refactored code to be more readable
