+++ src/components/Card.js
@@ -1,0 +1,7 @@
+const Card = ({cardImg}) => {
+    return(
+        <img id="card-img" alt="" src={cardImg}></img>
+    )
+}
+
+export default Card;
