+++ src/components/ReshuffleButton.js
@@ -1,0 +1,9 @@
+import "../App.css";
+
+const ReshuffleButton = ({reshuffle}) => {
+    return(
+        <button onClick={reshuffle} id="reshuffle">Reshuffle</button>
+    )
+}
+
+export default ReshuffleButton;
